+++ datadog_checks_base/datadog_checks/base/utils/db/utils.py
@@ -264,9 +264,6 @@ class DBMAsyncJob(object):
         Send a signal to cancel the job loop asynchronously.
         """
         self._cancel_event.set()
-        # after setting cancel event, wait for job loop to fully shutdown
-        if self._job_loop_future:
-            self._job_loop_future.result()
 
     def run_job_loop(self, tags):
         """